requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
et_xmlfile==2.0.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
